Refactor module structure.
