+++ crates/sui-genesis-builder/src/stardust/migration/verification/foundry.rs
@@ -1,13 +1,13 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use iota_sdk::types::block::output::AliasOutput;
+use iota_sdk::types::block::output::FoundryOutput;
 use sui_types::in_memory_storage::InMemoryStorage;
 
-use crate::stardust::migration::CreatedObjects;
+use super::created_objects::CreatedObjects;
 
-pub fn verify_alias_output(
-    output: &AliasOutput,
+pub fn verify_foundry_output(
+    output: &FoundryOutput,
     created_objects: &CreatedObjects,
     storage: &InMemoryStorage,
 ) -> anyhow::Result<()> {
